requirements.txt
google-adk>=1.26.0
